refactor: rename label to labels
