+++ internal/lib/labels/labels.go
@@ -1,4 +1,4 @@
-package label
+package labels
 
 import (
 	"errors"
